stock-fetcher/fetcher/requirements.txt
fastapi
psycopg2-binary
uvicorn
yfinance
